move run_umap_hnsw out of PARC class
