+++ README.md
@@ -55,6 +55,7 @@ We also show how to integrate UMAP with HNSW such that the embedding in UMAP is 
 
 ```
 import parc
+from parc.umap_hnsw import run_umap_hnsw
 import matplotlib.pyplot as plt
 from sklearn import datasets
 
@@ -78,7 +79,7 @@ plt.show()
 # Run umap on the HNSW knngraph already built in PARC (more time and memory efficient for large datasets)
 // Parc1.hnsw_index = p1.create_hnsw_index() // if you choose to visualize before running PARC clustering. then you need to include this line
 graph = Parc1.knngraph_full()
-X_umap = Parc1.run_umap_hnsw(X, graph)
+X_umap = run_umap_hnsw(X, graph)
 plt.scatter(X_umap[:, 0], X_umap[:, 1], c=Parc1.labels)
 plt.show()
